Add mypy config, clean views from unnecessary pydantic models, refactoring code by mypy
