+++ mypy.ini
@@ -1,0 +1,2 @@
+[mypy]
+allow_untyped_calls = true
