+++ service/api/views.py
@@ -1,9 +1,6 @@
-from typing import List, Optional, Sequence
-
 from fastapi import APIRouter, Depends, FastAPI, Request, Security
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 from fastapi.security.api_key import APIKey, APIKeyHeader, APIKeyQuery
-from pydantic import BaseModel
 
 from service.api.exceptions import (
     CredentialError,
@@ -13,31 +10,15 @@ from service.api.exceptions import (
 from service.log import app_logger
 
 from .config import config_env
+from .models import NotFoundError, RecoResponse, UnauthorizedError
 from .models_zoo import DumpModel
-
-
-class RecoResponse(BaseModel):
-    user_id: int
-    items: List[int]
-
-
-class NotFoundError(BaseModel):
-    error_key: str
-    error_message: str = "NotFound"
-    error_loc: Optional[Sequence[str]]
-
-
-class UnauthorizedError(BaseModel):
-    error_key: str
-    error_message: str = "Unauthorized"
-    error_loc: Optional[Sequence[str]]
-
 
 router = APIRouter()
 
 api_query = APIKeyQuery(name=config_env["API_KEY_NAME"], auto_error=False)
 api_header = APIKeyHeader(name=config_env["API_KEY_NAME"], auto_error=False)
 token_bearer = HTTPBearer(auto_error=False)
+
 models_zoo = {"model_1": DumpModel()}
